Split StaticHeader impl into separate module
This is a rough split, the next step is to induce better encapsulation.
